Unified timer handlingAs much as possible code is shared between both timer abstractions.
